requirements.txt
streamlit
google-generativeai
feedparser
yfinance
pandas_ta
pandas